// src/estate-favorites/new-estate-favorites-details.interface.ts
export interface NewEstateFavDetails {
  name: string;
  type: string;
  localisation: string;
  description: string;
  capacity: number;
  summary: string;
  amenities: string;
  image: string;
  piscine: boolean;
  restaurant: boolean;
  hammam: boolean;
  patrimoine: boolean;
  plage: boolean;
  randonnee: boolean;
  price: number;
}